Грамотное удаление домов

+++ app/api/endpoints/houses.py
@@ -42,7 +42,17 @@ async def get_house(house_service: HouseServiceDep, house_id: int):
     return house
 
 
-@houses_router.delete("/{house_id}", summary="Удаляет дом", status_code=status.HTTP_204_NO_CONTENT)
+@houses_router.delete(
+    "/{house_id}",
+    summary="Удаляет дом",
+    status_code=status.HTTP_204_NO_CONTENT,
+    responses={404: {'description': 'Дом не найден'}}
+)
 async def delete_house(house_service: HouseServiceDep, house_id: int):
+    house = await house_service.get_active_house(house_id=house_id)
+
+    if not house:
+        raise HTTPException(status_code=404, detail=f"Дом {house_id} не найден")
+
     await house_service.delete_house(house_id=house_id)
     return None
